Committer: root <root@kali.maxxaman>
 On branch main
 Your branch is up to date with 'origin/main'.

 Changes to be committed:
	modified:   __pycache__/config.cpython-39.pyc
	modified:   app/__pycache__/__init__.cpython-39.pyc
	modified:   app/__pycache__/errors.cpython-39.pyc
	modified:   app/__pycache__/forms.cpython-39.pyc
	modified:   app/__pycache__/models.cpython-39.pyc
	modified:   app/__pycache__/routes.cpython-39.pyc
	modified:   app/models.py
	new file:   logs/microblog.log
	new file:   tests.py

Message: I have added the ability to run a set of test on the app.
the test include checking password hashing funcionality user following and followed
It took sometime getting the tests to pass on the first run because I belive a bunch of recently added code was deleted
causing errors to run luckly it didnt take to long to find the missing code.